Major cleanup/reorganized
