+++ templates/index.html
@@ -48,7 +48,7 @@
             loading="lazy"
             allowfullscreen
             referrerpolicy="no-referrer-when-downgrade"
-            src="https://www.google.com/maps/embed/v1/place?key={{ .APIKey }}&q=ImpossibleLocation12345+NowhereEverThatExistsSurely">
+            src="https://www.google.com/maps/embed/v1/place?key={{ .GoogleMapsApiKey }}&q=ImpossibleLocation12345+NowhereEverThatExistsSurely">
         </iframe>
     </div>
 </body>
