+++ templates/searchInput.html
@@ -1,0 +1,16 @@
+<input 
+	class="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline mr-4" 
+	type="text" 
+	id="autocomplete" 
+	name="location" 
+	placeholder="Enter location" 
+	hx-get="/auto-complete/" 
+	hx-trigger="keyup changed delay:1000ms" 
+	hx-target="#suggestions" 
+	hx-indicator="#loading"
+	value="{{ . }}">
+<button 
+	class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline" 
+	type="submit">
+	Submit
+</button>
